fixes - code review notes
- name components
- fix style global cell
- split code to hooks
- split code to objects
