+++ src/Routes/SidebarRight/MemoryAndStorage/Bars.react.js
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { ResponsiveBar } from '@nivo/bar';
 import { Theme } from 'styles/colors';
+import useBars from './useBars';
 
 const Container = styled.div`
   font-size: 20px;
@@ -14,8 +15,14 @@ const Container = styled.div`
 
 // https://nivo.rocks/bar/ customization
 const Bars = ({ data, legend, colorScheme }) => {
-  const whiteColor = Theme.Styles.barColor.color;
-  const textColor = Theme.Styles.barColor.text;
+  const {
+    themePreferencesBar,
+    designBar,
+    fillBar,
+    axisLeftBar,
+    legendsBar,
+    axisBottomBar,
+  } = useBars();
 
   return (
     <Container>
@@ -23,26 +30,7 @@ const Bars = ({ data, legend, colorScheme }) => {
         data={data}
         keys={legend}
         indexBy="nodes"
-        theme={{
-          axis: {
-            ticks: {
-              line: {
-                stroke: textColor,
-              },
-              text: {
-                fontSize: 16,
-                marginRight: '10px',
-                fill: textColor,
-              },
-            },
-            legend: {
-              text: {
-                fontSize: 16,
-                fill: textColor,
-              },
-            },
-          },
-        }}
+        theme={themePreferencesBar}
         margin={{
           right: 0,
           bottom: 120,
@@ -56,97 +44,17 @@ const Bars = ({ data, legend, colorScheme }) => {
           colorScheme !== '' ? Theme.GRAPH_PALETTE : { scheme: colorScheme }
         }
         colorBy="id"
-        defs={[
-          {
-            id: 'dots',
-            type: 'patternDots',
-            background: 'inherit',
-            color: whiteColor,
-            size: 2,
-            padding: 3,
-            stagger: true,
-          },
-          {
-            id: 'lines',
-            type: 'patternLines',
-            background: 'inherit',
-            color: whiteColor,
-            rotation: -45,
-            lineWidth: 1,
-            spacing: 25,
-          },
-        ]}
-        fill={[
-          {
-            match: {
-              id: 'free',
-            },
-            id: 'dots',
-          },
-          {
-            match: {
-              id: 'reserved',
-            },
-            id: 'lines',
-          },
-        ]}
+        defs={designBar}
+        fill={fillBar}
         borderColor={Theme.COLOR.grey}
-        axisBottom={{
-          tickSize: 5,
-          tickPadding: 5,
-          tickRotation: 0,
-          legend: 'Size',
-          legendPosition: 'middle',
-          legendOffset: 50,
-        }}
-        axisLeft={{
-          format: v =>
-            v.length > 15 ? (
-              <tspan>
-                {`${v.substring(0, 15)}...`}
-                <title>{v}</title>
-              </tspan>
-            ) : (
-              v
-            ),
-          tickSize: 5,
-          tickPadding: 5,
-          tickRotation: 20,
-          legend: 'Nodes',
-          legendPosition: 'middle',
-          legendOffset: -50,
-          fontSize: 10,
-        }}
+        axisBottom={axisBottomBar}
+        axisLeft={axisLeftBar}
         labelSkipWidth={12}
         labelSkipHeight={12}
         animate
         motionStiffness={165}
         motionDamping={27}
-        legends={[
-          {
-            dataFrom: 'keys',
-            anchor: 'bottom',
-            direction: 'row',
-            justify: false,
-            translateX: 0,
-            translateY: 100,
-            itemsSpacing: 2,
-            itemWidth: 100,
-            itemHeight: 20,
-            itemDirection: 'left-to-right',
-            itemOpacity: 1,
-            symbolSize: 20,
-            itemTextColor: textColor,
-            effects: [
-              {
-                on: 'hover',
-                style: {
-                  itemOpacity: 0.85,
-                },
-              },
-            ],
-          },
-        ]}
+        legends={legendsBar}
       />
     </Container>
   );
