+++ src/Routes/Tables/Algorithms/Tabs/Versions/getVersionsColumns.react.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { CheckOutlined, DeleteOutlined } from '@ant-design/icons';
 import { Button, Modal, Tooltip, Typography, Tag } from 'antd';
 import Moment from 'react-moment';
-import { sorter } from 'utils/string';
+import { sorter } from 'utils/stringHelper';
 import { Ellipsis, FlexBox } from 'components/common';
 import { COLOR_PIPELINE_STATUS } from 'styles';
