+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as Table } from './Table';
+export { default as GlobalStyle } from './GlobalThemeStyle';
