+++ src/index.js
@@ -7,7 +7,7 @@ import { HashRouter as Router } from 'react-router-dom';
 import { ReusableProvider } from 'reusable';
 import { init } from 'actions/connection.action';
 import { selectors } from 'reducers';
-import GlobalStyle from './styles/GlobalStyle.styles';
+import { GlobalStyle } from './components';
 import Root from './Routes';
 import store from './store';
