+++ src/styles/GlobalStyle.styles.js
@@ -1,82 +1,52 @@
-import React, { useEffect, Suspense } from 'react';
-import { THEMES_NAMES } from 'const';
 import { createGlobalStyle } from 'styled-components';
 import { COLOR_LAYOUT, COLOR } from 'styles/colors';
-import { useSiteDarkMode } from 'hooks';
 
-// create in began styles
-if (localStorage.getItem('theme') === THEMES_NAMES.Dark) {
-  import('antd/dist/antd.dark.css');
-} else {
-  import('antd/dist/antd.css');
+const GlobalStyle = createGlobalStyle`
+* {
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
+    Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
 }
 
-// create in last styles to override styles
-const GlobalStyleDark = React.lazy(() => import('./GlobalStyleDark.styles'));
+.ant-tooltip-inner {
+  background-color: white;
+  color: black;
+}
 
-const GlobalStyle = () => {
-  const { isDarkMode, setTheme, themeName } = useSiteDarkMode();
-  useEffect(() => setTheme(themeName), []);
+.ant-layout-content{
+  overflow-x: hidden;
+}
 
-  const GlobalStyleHead = createGlobalStyle`
-  * {
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
-      Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  }
+.ant-layout-sider-light .ant-layout-sider-trigger {
+  border-right: 1px solid ${COLOR_LAYOUT.border};
+}
 
-  .ant-tooltip-inner {
-    background-color: white;
-    color: black;
-  }
-  
-  .ant-layout-content{
-    overflow-x: hidden;
-  }
+.ant-table-fixed-header .ant-table-scroll .ant-table-header {
+  overflow: hidden !important;
+  margin-bottom: 0px !important;
+}
 
-  .ant-layout-sider-light .ant-layout-sider-trigger {
-    border-right: 1px solid ${COLOR_LAYOUT.border};
-  }
+.ant-table table {
+  border-collapse: collapse;
+}
 
-  .ant-table-fixed-header .ant-table-scroll .ant-table-header {
-    overflow: hidden !important;
-    margin-bottom: 0px !important;
-  }
+::-webkit-scrollbar-track {
+  border: none;
+  background-color: none;
+}
+::-webkit-scrollbar {
+  width: 10px;
+  height: 10px;
+}
 
-  .ant-table table {
-    border-collapse: collapse;
-  }
+::-webkit-scrollbar-thumb {
+  border: 0.5px solid ${COLOR.grey};
+  background-color: ${COLOR_LAYOUT.border};
+}
 
-  ::-webkit-scrollbar-track {
-    border: none;
-    background-color: none;
-  }
-  ::-webkit-scrollbar {
-    width: 10px;
-    height: 10px;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    border: 0.5px solid ${COLOR.grey};
-    background-color: ${COLOR_LAYOUT.border};
-  }
-
-  .ant-layout {
-    background-color:  white;
-  }
+.ant-layout {
+  background-color:  white;
+}
 
 `;
-
-  return (
-    <>
-      <GlobalStyleHead isDarkMode={isDarkMode} />
-
-      {isDarkMode && (
-        <Suspense fallback="">
-          <GlobalStyleDark />
-        </Suspense>
-      )}
-    </>
-  );
-};
 
 export default GlobalStyle;
