+++ src/styles/index.js
@@ -1,4 +1,5 @@
 export { default as BottomPosition } from './BottomPosition.styles';
+
 export {
   Theme,
   COLOR,
@@ -12,9 +13,8 @@ export {
   COLOR_SERVICE,
   COLOR_TASK_STATUS,
 } from './colors';
-export { default as Display } from './Display.styles';
-// export { default as GlobalStyle } from './GlobalStyle.styles';
-// export { default as GlobalStyleDark } from './GlobalStyleDark.styles';
 
+export { default as Display } from './Display.styles';
+export { default as GlobalStyle } from './GlobalStyle.styles';
 export { default as mixins } from './mixins.styles';
 export { TabDrawerText, TabDrawer } from './Drawer.styles';
