+++ src/utils/stringHelper.js
